added exception handling in user and friendship classes

+++ src/main/java/com/assignment/socialnetworktool/service/impl/UserService.java
@@ -1,5 +1,8 @@
 package com.assignment.socialnetworktool.service.impl;
 
+import com.assignment.socialnetworktool.exception.FriendshipAlreadyExistsException;
+import com.assignment.socialnetworktool.exception.UserAlreadyExistsException;
+import com.assignment.socialnetworktool.exception.UserNotFoundException;
 import com.assignment.socialnetworktool.model.User;
 import com.assignment.socialnetworktool.repository.IUserRepository;
 import com.assignment.socialnetworktool.service.IUserService;
@@ -19,11 +22,17 @@ public class UserService implements IUserService {
 
     @Override
     public void addUser(User user) {
+        if(userRepository.exists(user.getUserId())){
+            throw new UserAlreadyExistsException("User already exists");
+        }
         userRepository.addUser(user);
     }
 
     @Override
     public void removeUser(String userId) {
+        if(!userRepository.exists(userId)){
+            throw new UserNotFoundException("User doesn't exists");
+        }
         userRepository.removeUser(userId);
     }
